feat: added item filter for receipts
Modify ReceiptFilters to include a new filter Item of type string
When this value is set, filter receipts which include any receipt item where name matches the requested value
Refactor queries in FindAllReceiptsForUserFilterByItem in order to store the query conditions into an array, instead of concatenating each one to the sql query.
This way is far more clear and sustainable than previous one.